+++ src/Stack/Types/Config.hs
@@ -58,6 +58,7 @@ module Stack.Types.Config
   -- ** ApplyGhcOptions
   ,ApplyGhcOptions(..)
   -- ** ConfigException
+  ,HpackExecutable(..)
   ,ConfigException(..)
   -- ** WhichSolverCmd
   ,WhichSolverCmd(..)
@@ -306,6 +307,8 @@ data Config =
          -- ^ How many concurrent jobs to run, defaults to number of capabilities
          ,configOverrideGccPath     :: !(Maybe (Path Abs File))
          -- ^ Optional gcc override path
+         ,configOverrideHpack       :: !HpackExecutable
+         -- ^ Use Hpack executable (overrides bundled Hpack)
          ,configExtraIncludeDirs    :: !(Set FilePath)
          -- ^ --extra-include-dirs arguments
          ,configExtraLibDirs        :: !(Set FilePath)
@@ -356,6 +359,11 @@ data Config =
          -- ^ Should we save Hackage credentials to a file?
          ,configRunner              :: !Runner
          }
+
+data HpackExecutable
+    = HpackBundled
+    | HpackCommand String
+    deriving (Show, Read, Eq, Ord)
 
 -- | Which packages do ghc-options on the command line apply to?
 data ApplyGhcOptions = AGOTargets -- ^ all local targets
@@ -705,6 +713,8 @@ data ConfigMonoid =
     -- ^ See: 'configExtraLibDirs'
     , configMonoidOverrideGccPath    :: !(First (Path Abs File))
     -- ^ Allow users to override the path to gcc
+    ,configMonoidOverrideHpack       :: !(First FilePath)
+    -- ^ Use Hpack executable (overrides bundled Hpack)
     ,configMonoidConcurrentTests     :: !(First Bool)
     -- ^ See: 'configConcurrentTests'
     ,configMonoidLocalBinPath        :: !(First FilePath)
@@ -789,6 +799,7 @@ parseConfigMonoidObject rootDir obj = do
     configMonoidExtraLibDirs <- fmap (Set.map (toFilePath rootDir FilePath.</>)) $
         obj ..:?  configMonoidExtraLibDirsName ..!= Set.empty
     configMonoidOverrideGccPath <- First <$> obj ..:? configMonoidOverrideGccPathName
+    configMonoidOverrideHpack <- First <$> obj ..:? configMonoidOverrideHpackName
     configMonoidConcurrentTests <- First <$> obj ..:? configMonoidConcurrentTestsName
     configMonoidLocalBinPath <- First <$> obj ..:? configMonoidLocalBinPathName
     configMonoidImageOpts <- jsonSubWarnings (obj ..:?  configMonoidImageOptsName ..!= mempty)
@@ -913,6 +924,9 @@ configMonoidExtraLibDirsName = "extra-lib-dirs"
 
 configMonoidOverrideGccPathName :: Text
 configMonoidOverrideGccPathName = "with-gcc"
+
+configMonoidOverrideHpackName :: Text
+configMonoidOverrideHpackName = "with-hpack"
 
 configMonoidConcurrentTestsName :: Text
 configMonoidConcurrentTestsName = "concurrent-tests"